refactor: migrate to schema input params
