+++ parser/src/schema/input_param.rs
@@ -1,12 +1,15 @@
+use rede_schema::InputParam as SchemaIP;
 use serde::Deserialize;
 
-/// Contains the different properties that can be defined for an input parameter.
 #[derive(Debug, Default, Deserialize, PartialEq)]
-pub struct InputParam {
-    /// Hint to provide to the user when asking for the input
-    pub hint: Option<String>,
-    /// Default value to use if the user does not provide any input
-    pub default: Option<String>,
+pub(crate) struct InputParam {
+    pub(crate) hint: Option<String>,
+}
+
+impl From<InputParam> for SchemaIP {
+    fn from(value: InputParam) -> Self {
+        Self { hint: value.hint }
+    }
 }
 
 #[cfg(test)]
@@ -17,14 +20,12 @@ mod test {
     fn deserialize() {
         let toml = r#"
             hint = "hint"
-            default = "default"
         "#;
         let input_param: InputParam = toml::from_str(toml).unwrap();
         assert_eq!(
             input_param,
             InputParam {
                 hint: Some("hint".to_string()),
-                default: Some("default".to_string()),
             }
         );
     }
